create league, join league, mock

+++ src/main/java/com/codecool/repository/UserRepository.java
@@ -1,12 +1,10 @@
 package com.codecool.repository;
 
-import com.codecool.model.League;
 import com.codecool.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     public boolean existsUserByUsername(String username);
